Add email view to contact users.
To be able to reply to offers made by the user we added the
functionality to send the user who made the offer an email. If the
owner wishes to. This required us to add an email attribute to the user
collection as well change how check if the item has any offers.